+++ R/varsel.R
@@ -495,6 +495,10 @@ parse_args_varsel <- function(refmodel, method, refit_prj, nterms_max,
   max_nv_possible <- count_terms_chosen(search_terms_unq) - 1L
   if (is.null(nterms_max)) {
     nterms_max <- 19
+    if (nterms_max < max_nv_possible &&
+        getOption("projpred.mssg_cut_search", TRUE)) {
+      message("Cutting off the search at size ", nterms_max, ".")
+    }
   }
   nterms_max <- min(max_nv_possible, nterms_max)